test_snake.py
def test_main():
    # Test case 1: Snake collides with the wall
    # The game should end when the snake collides with the wall
    assert main() == None

    # Test case 2: Snake collides with itself
    # The game should end when the snake collides with itself
    assert main() == None

    # Test case 3: Snake eats the food and grows
    # The snake should grow in length when it eats the food
    assert main() == None

    # Test case 4: Snake moves in the correct direction
    # The snake should move in the direction specified by the user's input
    assert main() == None

    # Test case 5: User quits the game
    # The game should end when the user quits
    assert main() == None